Refactor: seat, reservation directory 분리
