+++ src/main/kotlin/io/hhplus/concertreservationservice/domain/seat/SeatAlreadyReservedException.kt
@@ -1,3 +1,3 @@
-package io.hhplus.concertreservationservice.domain.reservation
+package io.hhplus.concertreservationservice.domain.seat
 
 class SeatAlreadyReservedException(message: String) : RuntimeException(message)
