+++ src/main/kotlin/io/hhplus/concertreservationservice/infrastructure/seat/JpaSeatRepository.kt
@@ -1,6 +1,6 @@
-package io.hhplus.concertreservationservice.infrastructure.reservation
+package io.hhplus.concertreservationservice.infrastructure.seat
 
-import io.hhplus.concertreservationservice.domain.reservation.Seat
+import io.hhplus.concertreservationservice.domain.seat.Seat
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
